# MedLog/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MedLog REST API",
    "version": "0.8.1.dev19+gff4bf8e.d20250714"
  },
  "paths": {
...
    "/api/user/me": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get Myself",
        "description": "Get account data from the current user",
        "operationId": "get_myself_api_user_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Update Myself",
        "description": "Update my user account data.",
        "operationId": "update_myself_api_user_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateByUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/user/me/password": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Set My Password",
        "description": "Set my password if i am a 'local' user. If my account was provisioned via an external OpenID Connect provider this does nothing except the return value will be `false`.",
        "operationId": "set_my_password_api_user_me_password_put",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_set_my_password_api_user_me_password_put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "title": "Response Set My Password Api User Me Password Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/user": {
      "post": {
        "tags": [
          "User Admin"
        ],
        "summary": "Create Local User",
        "description": "Creates a new user in the local user database. Need usermanager role",
        "operationId": "Create_local_user_api_user_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_password",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "The password for the created user. If non is defined the user will be created but not able to login until an admin user defines a password.",
              "title": "User Password"
            },
            "description": "The password for the created user. If non is defined the user will be created but not able to login until an admin user defines a password."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate",
                "description": "A json body with the user details"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "User Admin"
        ],
        "summary": "List Users",
        "description": "Get account data from a user by its id.  Need usermanager role",
        "operationId": "list_users_api_user_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incl_deactivated",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also list deactivated users.",
              "default": false,
              "title": "Incl Deactivated"
            },
            "description": "Also list deactivated users."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Specify the starting point for result sets/list",
              "default": 0,
              "title": "Offset"
            },
            "description": "Specify the starting point for result sets/list"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Specify the max amount of result items",
              "default": 100,
              "title": "Limit"
            },
            "description": "Specify the max amount of result items"
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "created_at",
                "email",
                "display_name",
                "id",
                "roles",
                "deactivated",
                "is_email_verified",
                "user_name"
              ],
              "type": "string",
              "description": "Order the result set by this attribute",
              "title": "Order By"
            },
            "description": "Order the result set by this attribute"
          },
          {
            "name": "order_desc",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Flip the sorting order",
              "default": false,
              "title": "Order Desc"
            },
            "description": "Flip the sorting order"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_User_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/{user_id}": {
      "get": {
        "tags": [
          "User Admin"
        ],
        "summary": "Get User",
        "description": "Get account data from a user by its id. Need usermanager role",
        "operationId": "get_user_api_user__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "User Admin"
        ],
        "summary": "Update User",
        "description": "Get account data from a user by its id. Need usermanager role",
        "operationId": "update_user_api_user__user_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateByAdmin",
                "description": "The user object with changed data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/{user_id}/password": {
      "put": {
        "tags": [
          "User Admin"
        ],
        "summary": "Set User Password",
        "description": "Set a local users password. If the user is provisioned via an external OpenID Connect provider the user will now be able to also login with basic login with this password.  Need usermanager role",
        "operationId": "set_user_password_api_user__user_id__password_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_set_user_password_api_user__user_id__password_put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_set_my_password_api_user_me_password_put": {
        "properties": {
          "old_password": {
            "type": "string",
            "title": "Old Password"
          },
          "new_password": {
            "type": "string",
            "title": "New Password"
          },
          "new_password_repeated": {
            "type": "string",
            "title": "New Password Repeated",
            "description": "For good measure we require the password twice to mitiage typos."
          }
        },
        "type": "object",
        "title": "Body_set_my_password_api_user_me_password_put"
      },
      "Body_set_user_password_api_user__user_id__password_put": {
        "properties": {
          "new_password": {
            "type": "string",
            "title": "New Password"
          },
          "new_password_repeated": {
            "type": "string",
            "title": "New Password Repeated",
            "description": "For good measure we require the password twice to mitiage typos."
          }
        },
        "type": "object",
        "required": [
          "new_password",
          "new_password_repeated"
        ],
        "title": "Body_set_user_password_api_user__user_id__password_put"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_User_": {
        "properties": {
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count",
            "description": "Total number of items in the database.",
            "examples": [
              300
            ]
          },
          "offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset",
            "description": "Starting position index of the returned items in the dataset.",
            "default": 0,
            "examples": [
              299
            ]
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Number of items returned in the response",
            "examples": [
              1
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "type": "array",
            "title": "Items",
            "description": "List of items returned in the response following given criteria"
          }
        },
        "type": "object",
        "required": [
          "count",
          "items"
        ],
        "title": "PaginatedResponse[User]"
      },
...
      "User": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 320
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "examples": [
              "clara@uni.wroc.pl",
              "titor@time.com"
            ]
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "examples": [
              "Clara Immerwahr",
              "John Titor"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles",
            "default": []
          },
          "deactivated": {
            "type": "boolean",
            "title": "Deactivated",
            "default": false
          },
          "is_email_verified": {
            "type": "boolean",
            "title": "Is Email Verified",
            "default": false
          },
          "user_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9.-_]+$",
            "title": "User Name",
            "examples": [
              "clara.immerwahr",
              "titor.extern.times"
            ]
          }
        },
        "type": "object",
        "required": [
          "user_name"
        ],
        "title": "User"
      },
      "UserCreate": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 320
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "examples": [
              "clara@uni.wroc.pl",
              "titor@time.com"
            ]
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "examples": [
              "Clara Immerwahr",
              "John Titor"
            ]
          },
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles",
            "default": []
          },
          "deactivated": {
            "type": "boolean",
            "title": "Deactivated",
            "default": false
          },
          "is_email_verified": {
            "type": "boolean",
            "title": "Is Email Verified",
            "default": false
          },
          "user_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9.-_]+$",
            "title": "User Name",
            "examples": [
              "clara.immerwahr",
              "titor.extern.times"
            ]
          }
        },
        "type": "object",
        "required": [
          "user_name"
        ],
        "title": "UserCreate"
      },
      "UserUpdateByAdmin": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 320
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "examples": [
              "clara@uni.wroc.pl",
              "titor@time.com"
            ]
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "examples": [
              "Clara Immerwahr",
              "John Titor"
            ]
          },
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles",
            "default": []
          },
          "deactivated": {
            "type": "boolean",
            "title": "Deactivated",
            "default": false
          },
          "is_email_verified": {
            "type": "boolean",
            "title": "Is Email Verified",
            "default": false
          }
        },
        "type": "object",
        "title": "UserUpdateByAdmin"
      },
      "UserUpdateByUser": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 320
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "examples": [
              "clara@uni.wroc.pl",
              "titor@time.com"
            ]
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "examples": [
              "Clara Immerwahr",
              "John Titor"
            ]
          }
        },
        "type": "object",
        "title": "UserUpdateByUser"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}